// src/app/api/model/moviePage.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MovieListItem } from './movieListItem';


export interface MoviePage { 
    page?: number;
    results?: Array<MovieListItem>;
    total_pages?: number;
    total_results?: number;
}