// src/app/api/model/tvSeriesPage.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TvSeriesListItem } from './tvSeriesListItem';


export interface TvSeriesPage { 
    page?: number;
    results?: Array<TvSeriesListItem>;
    total_pages?: number;
    total_results?: number;
}